chore: 去掉src/types
